+++ radicale/app/put.py
@@ -29,7 +29,7 @@ from radicale import app, httputils
 from radicale import item as radicale_item
 from radicale import pathutils, rights, storage, xmlutils
 from radicale.log import logger
-from radicale.hook.rabbitmq import QueueItem, QueueItemTypes
+from radicale.hook import HookNotificationItem, HookNotificationItemTypes
 
 MIMETYPE_TAGS = {value: key for key, value in xmlutils.MIMETYPES.items()}
 
@@ -195,7 +195,9 @@ class ApplicationPutMixin:
                     etag = self._storage.create_collection(
                         path, prepared_items, props).etag
                     for item in prepared_items:
-                        self._hook.notify(QueueItem(QueueItemTypes.UPSERT, item.serialize()))
+                        hook_notification_item = \
+                            HookNotificationItem(HookNotificationItemTypes.UPSERT, item.serialize())
+                        self._hook.notify(hook_notification_item)
                 except ValueError as e:
                     logger.warning(
                         "Bad PUT request on %r: %s", path, e, exc_info=True)
@@ -211,7 +213,9 @@ class ApplicationPutMixin:
                 href = posixpath.basename(pathutils.strip_path(path))
                 try:
                     etag = parent_item.upload(href, prepared_item).etag
-                    self._hook.notify(QueueItem(QueueItemTypes.UPSERT, prepared_item.serialize()))
+                    hook_notification_item = \
+                        HookNotificationItem(HookNotificationItemTypes.UPSERT, prepared_item.serialize())
+                    self._hook.notify(hook_notification_item)
                 except ValueError as e:
                     logger.warning(
                         "Bad PUT request on %r: %s", path, e, exc_info=True)
