+++ radicale/hook/none.py
@@ -2,5 +2,5 @@ from radicale import hook
 
 
 class Hook(hook.BaseHook):
-    def notify(self, content):
+    def notify(self, notification_item):
         """Notify nothing. Empty hook."""
